Refactorizacion del codigo del proyecto, creacion de carpetas que contienen cada una: un modelo, tipo y resolver de un modelo en especifico, organizacion del codigo del proyecto
